service and repositories commit

+++ smarthome-backend/src/main/java/com/example/smarthomebackend/Repositories/UserRepositorie.java
@@ -3,6 +3,6 @@ package com.example.smarthomebackend.Repositories;
 import com.example.smarthomebackend.Entites.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepositorie extends JpaRepository<User, Integer> {
+public interface UserRepositorie extends JpaRepository<User, Long> {
 
 }
